requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.6
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
ruff==0.12.10
